requirements.txt
openai
flask
python-binance